+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Go: Update `clustertest` to v5.0.0 & `cluster-standup-teardown` to v6.0.0.
+
 ## [6.4.0] - 2026-05-04
 
 ### Changed